feat(notifier): add notifier module
The notifier module is a simple wrapper around the notifyBC micro service instance that we plan to deploy.
Currently, only a limited set of functions are exposed (which fit the need of the notification system proposed):
-subscribe
-unsubscribe
-notify

When subscribe is called, the notifyBC api is called, creating an entry in it’s subscription data for a particular service and channel.
A “service” in this regard is really just a named notification group, so in this case we call it “opportunities”. The channel
specifies how the notification will be sent: email, sms, or inApp notifications. In our case, email in the primary use case.
Subscribing returns a promise with the subscription model as an argument.

Unsubscribe take a subscriptionId issues a request to unsubscribe the user from the notification.
The service will not delete the subscription model, but rather mark it as deleted by modifying it’s
state field to ‘deleted’.

The notify method will broadcast a message to all subscribed users to a service.

All of these methods return a Promise. It is up to the consumer to catch communication errors.