Subscribe user to opportunities notifications when notifyOpportunities set to true
When a user sets notifyOpportunities to true, we subscribe the user to the notifyBC micro service to receive emails when new opportunities are published. Communication with notifier is done before we save the record so that we can get the subscriptionId back from the notifiyBC and save it to the database. When the user choses to unsubscribe, this id will be used unsubscribe the user from the notifyBC database (this will mark the notifyBC subscription record state as ‘deleted’). A change in email address and the notifyOpporunities flag is set to true, it will update the subscription with the new email.