feat(opportunities): add email template (markdown/handlebars) and ability to no send notifications on update
The opportunities notification emails is now generated from body and subject markdown files located at `modules/opportunities/server/email_templates`. The `opportunity` object is passed to both of these files, so a user can combine custom text in addition to opportunity properties to create an email.

On `opportunity` updates, a `doNotNotify` flag is checked on the request. If this flag is not explicitly set, then no notifications will be sent (safety first).

A `lastPublished` property has been added to the `Opportunity` model. I use this to determine if the model have ever been published, which streamlines some of the UI when publishing for the first time.