Prompt user when navigating away from profile page
When a user modifies a form field on the user profile page, then navigates away from said page, the user will be prompted if they would like to continue or not. Continuing will discard all changes.

To implement this, the user model is copied on controller creation, avoiding changes to the Authenticated.user model without a save.