fix validations on edit profile form
Previously, an email address was required to edit a users profile field. An email is no longer required, however, if an email address is not entered, the user will not have the ability to check the notify or government flags (which require an email address). As per requirements, the first name and last name are no longer required.