add notify-bc service container to docker-compose
In order to run docker-compose, you’ll first need to build the notify-bc image from https://github.com/bcgov/MyGovBC-notification-server.git.

Notify-bc configuration is done by volume mapping a config file from this repo to the notify-bc container and setting the appropriate environment variables. This will get the dev environment running. With the current setup, the notify-bc container will not persist data to a database (rather to a file store).