Add update tasks
After the PR has been merged and deployed the notifyBC service will need to sync up with the devex database. Specifically, we need to subscribe users to the service that have the notifyOpportunities flag set.

Furthermore, there is `lastPublished` field added to the `Opportunities` model to check whether an opportunity has ever been published. This field must be updated in the database.

Run the following commands in the context of the production environment (i.e. production env variables for the database)

`gulp updateNotifyBC`

`gulp updateLastPublished`