Use db with notifyBC instead of file store
Updated docker-compose and added a datasources.development.js notifyBC configuration file.

Also, instead of whitelisting the hostname (which seemed to sometimes work, but not always), I am now whitelisting IPs from the docker subnet.

Now that a database is used on notifyBC, I needed to update the subscriptionOpportunitiesId field to use a string (id=“2342ad234”) instead of a number.