Notification email for scenario of unpublishing
If an opportunity is unpublished, users who are subscribed for updates on the opportunity should be notified. However, this may not mean that the creator is cancelling the opportunity - just taking it down temporarily. If the opportunity is unpublished, we can't provide a link to the opportunity (hence I haven't included a link).