Fixes issue #187:
Unsaved changes modal should not appear when deleting a proposal

Reproduced: Yes

Root Cause: Proposal is not updated when calling `$remove()` on the resource.

Solution: Function added `unsetProposal()` which sets controller's proposal and internal pristine proposal to empty object. Thus when `$locationChangeStartUnbind()` is called both the controllers proposal and pristine proposal are the same.